rpcclient: move streaming cache tests
